+++ coloraide/spaces/cam02.py
@@ -278,6 +278,16 @@ class CAM02JMh(LCh):
         Channel("h", 0.0, 360.0, flags=FLG_ANGLE)
     )
 
+    def get_lightness_name(self) -> str:
+        """Get lightness name."""
+
+        return "j"
+
+    def radial_name(self) -> str:
+        """Get radial name."""
+
+        return "m"
+
     def normalize(self, coords: Vector) -> Vector:
         """Normalize."""