+++ coloraide/spaces/cam16.py
@@ -386,6 +386,16 @@ class CAM16JMh(LCh):
         Channel("h", 0.0, 360.0, flags=FLG_ANGLE)
     )
 
+    def lightness_name(self) -> str:
+        """Get lightness name."""
+
+        return "j"
+
+    def radial_name(self) -> str:
+        """Get radial name."""
+
+        return "m"
+
     def normalize(self, coords: Vector) -> Vector:
         """Normalize."""