+++ coloraide/spaces/cmyk.py
@@ -57,6 +57,7 @@ class CMYK(Space):
         "black": 'k'
     }
     WHITE = WHITES['2deg']['D65']
+    SUBTRACTIVE = True
 
     def is_achromatic(self, coords: Vector) -> bool:
         """Test if color is achromatic."""